+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 ### Added
 - execsnoop related routes
   - get process status utilizing pgrep and psutil
-  
+  - load all execsnoop process to objects
+
 - redis integration
 - docker testing with redis container
-- configuration pytests
+- pytests
+  - configuration
+  - execsnoop instances manipulation
+
+- serialize all existing execsnoop instances
 
 [Unreleased]: https://github.com/tomas321/sneakctl_server/compare/develop